Fixed several deprecation warnings, by product of the Julia version moving from v0.3 to v0.6
The keywords used to define types have changed (#19157, #20418) of JuliaLang/julia.
"abstract changes to abstract type ... end", fix for the same.

The typealias keyword is deprecated, and should be replaced with Vector{T} = Array{T,1} or a const assignment (#20500) of JuliaLang/julia.
Fix for the same

Changed Base.(operator) to Base.operator to fix deprecation warnings. Note, this should be replaced in the future
as one no longer defines methods for .* etcetera due to (#17623) of JuliaLang/julia.